split ast
