+++ ast/ast.go
@@ -2,10 +2,6 @@ package ast
 
 import (
 	"bytes"
-	"fmt"
-	"strings"
-
-	"github.com/flipez/rocket-lang/token"
 )
 
 func (p *Program) String() string {
@@ -14,79 +10,6 @@ func (p *Program) String() string {
 	for _, s := range p.Statements {
 		out.WriteString(s.String())
 	}
-
-	return out.String()
-}
-
-type HashLiteral struct {
-	Token token.Token
-	Pairs map[Expression]Expression
-}
-
-func (hl *HashLiteral) expressionNode()      {}
-func (hl *HashLiteral) TokenLiteral() string { return hl.Token.Literal }
-func (hl *HashLiteral) String() string {
-	var out bytes.Buffer
-
-	pairs := []string{}
-	for key, value := range hl.Pairs {
-		pairs = append(pairs, key.String()+":"+value.String())
-	}
-
-	out.WriteString("{")
-	out.WriteString(strings.Join(pairs, ", "))
-	out.WriteString("}")
-
-	return out.String()
-}
-
-type IndexExpression struct {
-	Token token.Token
-	Left  Expression
-	Index Expression
-}
-
-func (ie *IndexExpression) expressionNode()      {}
-func (ie *IndexExpression) TokenLiteral() string { return ie.Token.Literal }
-func (ie *IndexExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString("(")
-	out.WriteString(ie.Left.String())
-	out.WriteString("[")
-	out.WriteString(ie.Index.String())
-	out.WriteString("])")
-
-	return out.String()
-}
-
-type StringLiteral struct {
-	Token token.Token
-	Value string
-}
-
-func (sl *StringLiteral) expressionNode()      {}
-func (sl *StringLiteral) TokenLiteral() string { return sl.Token.Literal }
-func (sl *StringLiteral) String() string       { return sl.Token.Literal }
-
-type ArrayLiteral struct {
-	Token    token.Token
-	Elements []Expression
-}
-
-func (al *ArrayLiteral) expressionNode()      {}
-func (al *ArrayLiteral) TokenLiteral() string { return al.Token.Literal }
-func (al *ArrayLiteral) String() string {
-	var out bytes.Buffer
-
-	elements := []string{}
-	for _, el := range al.Elements {
-		elements = append(elements, el.String())
-	}
-
-	out.WriteString("[")
-	out.WriteString(strings.Join(elements, ", "))
-	out.WriteString("]")
 
 	return out.String()
 }
@@ -116,287 +39,4 @@ func (p *Program) TokenLiteral() string {
 	} else {
 		return ""
 	}
-}
-
-type ReturnStatement struct {
-	Token       token.Token
-	ReturnValue Expression
-}
-
-func (rs *ReturnStatement) statementNode()       {}
-func (rs *ReturnStatement) TokenLiteral() string { return rs.Token.Literal }
-func (rs *ReturnStatement) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(rs.TokenLiteral() + " ")
-
-	if rs.ReturnValue != nil {
-		out.WriteString(rs.ReturnValue.String())
-	}
-
-	out.WriteString(";")
-
-	return out.String()
-}
-
-type ExpressionStatement struct {
-	Token      token.Token
-	Expression Expression
-}
-
-func (es *ExpressionStatement) statementNode()       {}
-func (es *ExpressionStatement) TokenLiteral() string { return es.Token.Literal }
-func (es *ExpressionStatement) String() string {
-	if es.Expression != nil {
-		return es.Expression.String()
-	}
-
-	return ""
-}
-
-type Identifier struct {
-	Token token.Token // the token.IDENT token
-	Value string
-}
-
-func (i *Identifier) expressionNode()      {}
-func (i *Identifier) TokenLiteral() string { return i.Token.Literal }
-func (i *Identifier) String() string       { return i.Value }
-
-type IntegerLiteral struct {
-	Token token.Token
-	Value int64
-}
-
-func (il *IntegerLiteral) expressionNode()      {}
-func (il *IntegerLiteral) TokenLiteral() string { return il.Token.Literal }
-func (il *IntegerLiteral) String() string       { return il.Token.Literal }
-
-type FloatLiteral struct {
-	Token token.Token
-	Value float64
-}
-
-func (fl *FloatLiteral) expressionNode()      {}
-func (fl *FloatLiteral) TokenLiteral() string { return fl.Token.Literal }
-func (fl *FloatLiteral) String() string       { return fl.Token.Literal }
-
-type PrefixExpression struct {
-	Token    token.Token
-	Operator string
-	Right    Expression
-}
-
-func (pe *PrefixExpression) expressionNode()      {}
-func (pe *PrefixExpression) TokenLiteral() string { return pe.Token.Literal }
-func (pe *PrefixExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString("(")
-	out.WriteString(pe.Operator)
-	out.WriteString(pe.Right.String())
-	out.WriteString(")")
-
-	return out.String()
-}
-
-type InfixExpression struct {
-	Token    token.Token
-	Left     Expression
-	Operator string
-	Right    Expression
-}
-
-func (ie *InfixExpression) expressionNode()      {}
-func (ie *InfixExpression) TokenLiteral() string { return ie.Token.Literal }
-func (ie *InfixExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString("(")
-	out.WriteString(ie.Left.String())
-	out.WriteString(" " + ie.Operator + " ")
-	out.WriteString(ie.Right.String())
-	out.WriteString(")")
-
-	return out.String()
-}
-
-type Boolean struct {
-	Token token.Token
-	Value bool
-}
-
-func (b *Boolean) expressionNode()      {}
-func (b *Boolean) TokenLiteral() string { return b.Token.Literal }
-func (b *Boolean) String() string       { return b.Token.Literal }
-
-type IfExpression struct {
-	Token       token.Token // the if token
-	Condition   Expression
-	Consequence *BlockStatement
-	Alternative *BlockStatement
-}
-
-func (ie *IfExpression) expressionNode()      {}
-func (ie *IfExpression) TokenLiteral() string { return ie.Token.Literal }
-func (ie *IfExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString("if ")
-	out.WriteString(ie.Condition.String())
-	out.WriteString(" ")
-	out.WriteString(ie.Consequence.String())
-
-	if ie.Alternative != nil {
-		out.WriteString(" else ")
-		out.WriteString(ie.Alternative.String())
-	}
-	out.WriteString(" end")
-
-	return out.String()
-}
-
-type BlockStatement struct {
-	Token      token.Token // the { token
-	Statements []Statement
-}
-
-func (bs *BlockStatement) statementNode()       {}
-func (bs *BlockStatement) TokenLiteral() string { return bs.Token.Literal }
-func (bs *BlockStatement) String() string {
-	var out bytes.Buffer
-
-	for _, s := range bs.Statements {
-		out.WriteString(s.String())
-	}
-
-	return out.String()
-}
-
-type FunctionLiteral struct {
-	Token      token.Token
-	Name       string
-	Parameters []*Identifier
-	Body       *BlockStatement
-}
-
-func (fl *FunctionLiteral) expressionNode()      {}
-func (fl *FunctionLiteral) TokenLiteral() string { return fl.Token.Literal }
-func (fl *FunctionLiteral) String() string {
-	var out bytes.Buffer
-
-	params := []string{}
-	for _, p := range fl.Parameters {
-		params = append(params, p.String())
-	}
-
-	out.WriteString(fl.TokenLiteral())
-	out.WriteString("(")
-	out.WriteString(strings.Join(params, ", "))
-	out.WriteString(") ")
-	out.WriteString(fl.Body.String())
-
-	return out.String()
-}
-
-type CallExpression struct {
-	Token     token.Token // The ( token
-	Callable  Expression
-	Arguments []Expression
-}
-
-func (ce *CallExpression) expressionNode()      {}
-func (ce *CallExpression) TokenLiteral() string { return ce.Token.Literal }
-func (ce *CallExpression) String() string {
-	var out bytes.Buffer
-
-	args := []string{}
-	for _, a := range ce.Arguments {
-		args = append(args, a.String())
-	}
-
-	out.WriteString(ce.Callable.String())
-	out.WriteString("(")
-	out.WriteString(strings.Join(args, ", "))
-	out.WriteString(")")
-
-	return out.String()
-}
-
-type ObjectCallExpression struct {
-	Token  token.Token
-	Object Expression
-	Call   Expression
-}
-
-func (oce *ObjectCallExpression) expressionNode() {}
-
-func (oce *ObjectCallExpression) TokenLiteral() string {
-	return oce.Token.Literal
-}
-
-func (oce *ObjectCallExpression) String() string {
-	var out bytes.Buffer
-	out.WriteString(oce.Object.String())
-	out.WriteString(".")
-	out.WriteString(oce.Call.String())
-
-	return out.String()
-}
-
-type ForeachStatement struct {
-	Token token.Token
-	Index string
-	Ident string
-	Value Expression
-	Body  *BlockStatement
-}
-
-func (fes *ForeachStatement) expressionNode()      {}
-func (fes *ForeachStatement) TokenLiteral() string { return fes.Token.Literal }
-func (fes *ForeachStatement) String() string {
-	var out bytes.Buffer
-	out.WriteString("foreach ")
-	out.WriteString(fes.Ident)
-	out.WriteString(" ")
-	out.WriteString(fes.Value.String())
-	out.WriteString(fes.Body.String())
-	return out.String()
-}
-
-type AssignStatement struct {
-	Token token.Token
-	Name  *Identifier
-	Value Expression
-}
-
-func (as *AssignStatement) expressionNode()      {}
-func (as *AssignStatement) statementNode()       {}
-func (as *AssignStatement) TokenLiteral() string { return as.Token.Literal }
-func (as *AssignStatement) String() string {
-	var out bytes.Buffer
-	out.WriteString(as.Name.String())
-	out.WriteString(" = ")
-	out.WriteString(as.Value.String())
-	return out.String()
-}
-
-type ImportExpression struct {
-	Token token.Token
-	Name  Expression
-}
-
-func (ie *ImportExpression) expressionNode() {}
-
-func (ie *ImportExpression) TokenLiteral() string { return ie.Token.Literal }
-
-func (ie *ImportExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(ie.TokenLiteral())
-	out.WriteString("(")
-	out.WriteString(fmt.Sprintf("\"%s\"", ie.Name))
-	out.WriteString(")")
-
-	return out.String()
 }
