+++ ast/call_expression.go
@@ -1,0 +1,32 @@
+package ast
+
+import (
+	"bytes"
+	"strings"
+
+	"github.com/flipez/rocket-lang/token"
+)
+
+type CallExpression struct {
+	Token     token.Token // The ( token
+	Callable  Expression
+	Arguments []Expression
+}
+
+func (ce *CallExpression) expressionNode()      {}
+func (ce *CallExpression) TokenLiteral() string { return ce.Token.Literal }
+func (ce *CallExpression) String() string {
+	var out bytes.Buffer
+
+	args := []string{}
+	for _, a := range ce.Arguments {
+		args = append(args, a.String())
+	}
+
+	out.WriteString(ce.Callable.String())
+	out.WriteString("(")
+	out.WriteString(strings.Join(args, ", "))
+	out.WriteString(")")
+
+	return out.String()
+}
