+++ ast/float_literal.go
@@ -1,0 +1,14 @@
+package ast
+
+import (
+	"github.com/flipez/rocket-lang/token"
+)
+
+type FloatLiteral struct {
+	Token token.Token
+	Value float64
+}
+
+func (fl *FloatLiteral) expressionNode()      {}
+func (fl *FloatLiteral) TokenLiteral() string { return fl.Token.Literal }
+func (fl *FloatLiteral) String() string       { return fl.Token.Literal }
