+++ ast/identifier.go
@@ -1,0 +1,14 @@
+package ast
+
+import (
+	"github.com/flipez/rocket-lang/token"
+)
+
+type Identifier struct {
+	Token token.Token // the token.IDENT token
+	Value string
+}
+
+func (i *Identifier) expressionNode()      {}
+func (i *Identifier) TokenLiteral() string { return i.Token.Literal }
+func (i *Identifier) String() string       { return i.Value }
