+++ ast/if_expression.go
@@ -1,0 +1,33 @@
+package ast
+
+import (
+	"bytes"
+
+	"github.com/flipez/rocket-lang/token"
+)
+
+type IfExpression struct {
+	Token       token.Token // the if token
+	Condition   Expression
+	Consequence *BlockStatement
+	Alternative *BlockStatement
+}
+
+func (ie *IfExpression) expressionNode()      {}
+func (ie *IfExpression) TokenLiteral() string { return ie.Token.Literal }
+func (ie *IfExpression) String() string {
+	var out bytes.Buffer
+
+	out.WriteString("if ")
+	out.WriteString(ie.Condition.String())
+	out.WriteString(" ")
+	out.WriteString(ie.Consequence.String())
+
+	if ie.Alternative != nil {
+		out.WriteString(" else ")
+		out.WriteString(ie.Alternative.String())
+	}
+	out.WriteString(" end")
+
+	return out.String()
+}
